Move more files
Move VoxelNormalsPalette to TS
