+++ OpenRA.Mods.Cnc/Traits/Render/WithDeliveryAnimation.cs
@@ -8,9 +8,10 @@
  */
 #endregion
 
+using OpenRA.Mods.RA.Render;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Render
+namespace OpenRA.Mods.Cnc.Traits
 {
 	[Desc("Building animation to play when ProductionAirdrop is used to deliver units.")]
 	public class WithDeliveryAnimationInfo : ITraitInfo, Requires<RenderBuildingInfo>
