+++ OpenRA.Mods.RA/AI/Squad.cs
@@ -10,8 +10,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using OpenRA.Mods.RA;
-using OpenRA.Mods.RA.Traits;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Support;
 using OpenRA.Traits;
