+++ OpenRA.Mods.TS/Activities/VoxelHarvesterDockSequence.cs
@@ -11,12 +11,11 @@
 using System;
 using System.Collections.Generic;
 using OpenRA.Mods.RA.Activities;
-using OpenRA.Mods.RA.Move;
-using OpenRA.Mods.RA.Render;
 using OpenRA.Mods.RA.Traits;
+using OpenRA.Mods.TS.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.TS
+namespace OpenRA.Mods.TS.Activities
 {
 	public class VoxelHarvesterDockSequence : Activity
 	{
