+++ OpenRA.Mods.TS/OpenRA.Mods.TS.csproj
@@ -58,8 +58,13 @@
     <Compile Include="Activities\VoxelHarvesterDockSequence.cs" />
     <Compile Include="SpriteLoaders\TmpTSLoader.cs" />
     <Compile Include="Traits\Buildings\TiberianSunRefinery.cs" />
+    <Compile Include="Traits\Render\RenderVoxels.cs" />
+    <Compile Include="Traits\Render\WithVoxelBarrel.cs" />
+    <Compile Include="Traits\Render\WithVoxelBody.cs" />
+    <Compile Include="Traits\Render\WithVoxelTurret.cs" />
     <Compile Include="Traits\Render\WithVoxelWalkerBody.cs" />
     <Compile Include="Traits\Render\WithVoxelUnloadBody.cs" />
+    <Compile Include="Traits\World\VoxelNormalsPalette.cs" />
     <Compile Include="Traits\World\ShroudPalette.cs" />
     <Compile Include="UtilityCommands\LegacyTilesetImporter.cs" />
   </ItemGroup>
